+++ lib/app/modules/detail_buku/views/web_screen.dart
@@ -274,7 +274,7 @@ class Web_Screen extends GetView<DetailBukuController> {
                                   : 'Tidak Ditemukan')),
                               DataCell(Text(comment.ulasan ?? '')),
                               DataCell(RatingBarIndicator(
-                                rating: comment.rating!,
+                                rating: comment.rating!.toDouble(),
                                 itemBuilder: (context, index) => const Icon(
                                   Icons.star,
                                   color: Colors.amber,
